+++ x/evidence/keeper/grpc_query.go
@@ -68,18 +68,12 @@ func (k Querier) AllEvidence(ctx context.Context, req *types.QueryAllEvidenceReq
 		return nil, status.Errorf(codes.InvalidArgument, "empty request")
 	}
 
-	var evidence []*codectypes.Any
-	_, pageRes, err := query.CollectionFilteredPaginate(ctx, k.k.Evidences, req.Pagination, func(_ []byte, value exported.Evidence) (include bool, err error) {
-		evidenceAny, err := codectypes.NewAnyWithValue(value)
-		if err != nil {
-			return false, err
-		}
-		evidence = append(evidence, evidenceAny)
-		return false, nil // we don't include results because we're appending them
+	evidences, pageRes, err := query.CollectionPaginate(ctx, k.k.Evidences, req.Pagination, func(_ []byte, value exported.Evidence) (*codectypes.Any, error) {
+		return codectypes.NewAnyWithValue(value)
 	})
 	if err != nil {
 		return nil, err
 	}
 
-	return &types.QueryAllEvidenceResponse{Evidence: evidence, Pagination: pageRes}, nil
+	return &types.QueryAllEvidenceResponse{Evidence: evidences, Pagination: pageRes}, nil
 }